#8 Refactor
