+++ src/InvScanPro/AppShell.xaml
@@ -3,12 +3,13 @@
     x:Class="InvScanPro.AppShell"
     xmlns="http://schemas.microsoft.com/dotnet/2021/maui"
     xmlns:x="http://schemas.microsoft.com/winfx/2009/xaml"
+    xmlns:v="clr-namespace:InvScanPro.Views"
     xmlns:local="clr-namespace:InvScanPro"
     Shell.FlyoutBehavior="Disabled">
 
     <ShellContent
         Title="{StaticResource Label_0012}"
-        ContentTemplate="{DataTemplate local:MainPage}"
+        ContentTemplate="{DataTemplate v:MainPage}"
         Route="MainPage" />
 
 </Shell>
