+++ src/InvScanPro/ViewModels/MainViewModel.cs
@@ -1,5 +1,6 @@
 ﻿using CommunityToolkit.Mvvm.ComponentModel;
 using CommunityToolkit.Mvvm.Input;
+using InvScanPro.Views;
 
 namespace InvScanPro.ViewModels;
