+++ src/InvScanPro/Views/DatePage.xaml
@@ -1,37 +1,36 @@
 <?xml version="1.0" encoding="utf-8" ?>
 <ContentPage xmlns="http://schemas.microsoft.com/dotnet/2021/maui"
              xmlns:x="http://schemas.microsoft.com/winfx/2009/xaml"
-             xmlns:vm="clr-namespace:InvScanPro.ViewModels"
-             x:Class="InvScanPro.LocationPage"
-             x:DataType="vm:LocationViewModel"
-             Title="{StaticResource Label_0014}">
+             xmlns:vm ="clr-namespace:InvScanPro.ViewModels"
+             x:Class="InvScanPro.Views.DatePage"
+             x:DataType="vm:DateViewModel"
+             Title="{StaticResource Label_0013}">
     <ScrollView>
         <VerticalStackLayout
             Spacing="30"
             Padding="30,0"
             VerticalOptions="Start">
 
-            <Label Text="{StaticResource Label_0010}"
+            <Label Text="{StaticResource Label_0005}"
                    HeightRequest="60"
                    WidthRequest="215"
                    Margin="0,30,0,0"
                    FontSize="18"
                    HorizontalOptions="Center"/>
 
-            <Entry Placeholder="{StaticResource Label_0015}"
-                   Text="{Binding Location}"
-                   HeightRequest="60"
-                   WidthRequest="200"
-                   FontSize="18"
-                   HorizontalOptions="Center"/>
+            <DatePicker Date="{Binding Date}"
+                        HeightRequest="60"
+                        WidthRequest="200"
+                        FontSize="18"
+                        HorizontalOptions="Center"/>
 
             <Button Text="{StaticResource Label_0006}"
                     HeightRequest="60"
                     WidthRequest="200" 
                     FontSize="18"
-                    SemanticProperties.Hint="{StaticResource Label_0011}"
+                    SemanticProperties.Hint="{StaticResource Label_0007}"
                     HorizontalOptions="Center"
-                    Command="{Binding NavigateToGeneralPageCommand}"/>
+                    Command="{Binding NavigateToLocationPageCommand}"/>
 
             <Button Text="{StaticResource Label_0008}"
                     HeightRequest="60"
