+++ src/InvScanPro/Views/LocationPage.xaml.cs
@@ -1,6 +1,6 @@
 using InvScanPro.ViewModels;
 
-namespace InvScanPro;
+namespace InvScanPro.Views;
 
 public partial class LocationPage : ContentPage
 {
